trace: cheri: function graph tracing
Support function graph tracing for cheri.

Function graph tracing adds hooks to the start and to the exit of a traced
function. Here's a basic summary of the mechanism:

start of the traced function
   jump into trampoline (ftrace_caller, arch/riscv/kernel/mcount-dyn.S)
      jump into graph handler
         save traced function's ra
         overwrite ra it with graph return handler
         call graph "start" functions
   go back to traced function

traced function exits
   "return" into graph return handler
   call graph "end" functions
   restore the original ra
   return to caller of the traced function

The graph handler (ftrace_graph_func) saves ra in struct ftrace_ret_stack's ret
component. We have to save and restore the complete register, ret must be
converted to a capability.

While the traced function runs, ra is replaced with a capability for the graph
return handler (return_to_handler) that allows executing this function.

Each task has an array of ftrace_ret_stack entries. current->ret_stack is a
pointer to this "return stack", it must become a capability.

Many places throughout the tracing code assume that a struct ftrace_ret_stack
can be saved at any current->ret_stack[i] position of the return stack. We have
to bite the bullet and make the return stack an array of uintptr_t. This wastes
some stack space when standard data fields are stored onto the stack. On the
other hand, the entries are short-lived, not passed to user space and the
recursion depth for function tracing is usually not too deep (or the trace will
become unreadable). Therefore, capability-aligned entries should be ok.

+++ include/linux/ftrace.h
@@ -44,7 +44,7 @@ struct dyn_ftrace;
 char *arch_ftrace_match_adjust(char *str, const char *search);
 
 #ifdef CONFIG_HAVE_FUNCTION_GRAPH_FREGS
-unsigned long ftrace_return_to_handler(struct ftrace_regs *fregs);
+uintptr_t ftrace_return_to_handler(struct ftrace_regs *fregs);
 #else
 unsigned long ftrace_return_to_handler(unsigned long frame_pointer);
 #endif
@@ -1265,12 +1265,12 @@ void *fgraph_retrieve_parent_data(int idx, int *size_bytes, int depth);
  * Used in struct thread_info
  */
 struct ftrace_ret_stack {
-	unsigned long ret;
+	uintptr_t ret;
 	unsigned long func;
 #ifdef HAVE_FUNCTION_GRAPH_FP_TEST
 	unsigned long fp;
 #endif
-	unsigned long *retp;
+	uintptr_t *retp;
 };
 
 /*
@@ -1281,12 +1281,12 @@ struct ftrace_ret_stack {
 extern void return_to_handler(void);
 
 extern int
-function_graph_enter_regs(unsigned long ret, unsigned long func,
-			  unsigned long frame_pointer, unsigned long *retp,
+function_graph_enter_regs(uintptr_t ret, unsigned long func,
+			  unsigned long frame_pointer, uintptr_t *retp,
 			  struct ftrace_regs *fregs);
 
-static inline int function_graph_enter(unsigned long ret, unsigned long func,
-				       unsigned long fp, unsigned long *retp)
+static inline int function_graph_enter(uintptr_t ret, unsigned long func,
+				       unsigned long fp, uintptr_t *retp)
 {
 	return function_graph_enter_regs(ret, func, fp, retp, NULL);
 }
@@ -1296,8 +1296,8 @@ ftrace_graph_get_ret_stack(struct task_struct *task, int skip);
 unsigned long ftrace_graph_top_ret_addr(struct task_struct *task);
 
 unsigned long ftrace_graph_ret_addr(struct task_struct *task, int *idx,
-				    unsigned long ret, unsigned long *retp);
-unsigned long *fgraph_get_task_var(struct fgraph_ops *gops);
+				    unsigned long ret, uintptr_t *retp);
+uintptr_t *fgraph_get_task_var(struct fgraph_ops *gops);
 
 /*
  * Sometimes we don't want to trace a function with the function
