+++ include/linux/sched.h
@@ -1476,7 +1476,7 @@ struct task_struct {
 	int				curr_ret_depth;
 
 	/* Stack of return addresses for return function tracing: */
-	unsigned long			*ret_stack;
+	uintptr_t			*ret_stack;
 
 	/* Timestamp for last schedule: */
 	unsigned long long		ftrace_timestamp;
